fix: handle kaleido browser detection better
Importing the Plotly kaleido scope for image defaults triggers a search
for a Chromium based browser, that, when it fails, stops all cli
commands from loading. Handle this error better by replacing the
defaults with `None` values and provide a more user-friendly version
of the error feedback when trying to export an image.