fix: Correct image format selection
When setting an output filename, figure.write_image is passed an open
fileobject and so it'd never select a format based on the filename
extension. Explicitly set a format when none has been seleceted, and
don't try to second-guess the user when a format has been passed in
that conflicts with the filename.